requirements.txt
certifi==2021.5.30
chardet==4.0.0
ibm-cloud-sdk-core==3.10.0
ibm-watson==5.2.0
idna==2.10
PyJWT==2.1.0
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
urllib3==1.26.5
websocket-client==0.48.0
